requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.1.1
werkzeug==2.3.7
boto3==1.28.62
python-dotenv==1.0.0
gunicorn==21.2.0
flask-wtf==1.1.1
email-validator==2.0.0
pytest 